ISSUE-1240 Use JMAPExtensionConfiguration to handle email recovery config
